+++ CHANGELOG.md
@@ -13,6 +13,11 @@
 - [`LanguageService.resolveRenameLocation`](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService#resolverenamelocation) should ignore file and category links
 - The outputs of [`wikiparse.Linter.prototype.codemirror`](https://github.com/bhsd-harry/wikiparser-node/wiki/wikiparse.Linter#codemirror) and [`wikiparse.Linter.prototype.monaco`](https://github.com/bhsd-harry/wikiparser-node/wiki/wikiparse.Linter#monaco) now strictly follow the [CodeMirror](https://codemirror.net/docs/ref/#lint.Diagnostic) and [Monaco Editor](https://microsoft.github.io/monaco-editor/docs.html#interfaces/editor.IMarkerData.html) specifications, respectively
 
+**Changed**
+
+- The return values of [`Token.prototype.caretPositionfromIndex`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token#caretpositionfromindex) and its related methods are not different if the index is at the boundary between two tokens
+- [`Token.prototype.elementFromIndex`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token#elementfromindex) and its related methods now return [`Token`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token) instead of [`AstNode`](https://github.com/bhsd-harry/wikiparser-node/wiki/AstNode)
+
 ## v1.16.1
 
 *2025-02-10*